update shunthing yard to push to the que To the client. update ConnectControlClientCommand to push information from the que, and make some functions to variables and var

+++ Var.cpp
@@ -19,7 +19,7 @@ const string &Var::GetSim() const {
 
 }
 
-void Var::setValue(int value) {
+void Var::setValue(double value) {
     Var::value = value;
 }
 //
